+++ bootique-jersey-jakarta-client-junit5-wiremock/src/test/java/io/bootique/jersey/client/junit5/wiremock/WireMockTester_WithBasePathIT.java
@@ -36,7 +36,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class WireMockTester_WithBasePathIT {
 
     @BQTestTool
-    static final WireMockTester tester = WireMockTester.tester("https://github.com/bootique");
+    static final WireMockTester tester = WireMockTester.create("https://github.com/bootique");
 
     @BQApp(skipRun = true)
     static final BQRuntime app = Bootique.app()